+++ flutter_sound/example/lib/recordToStream/record_to_stream_example.dart
@@ -19,9 +19,7 @@
 
 import 'dart:async';
 import 'dart:io';
-import 'dart:typed_data';
 import 'package:audio_session/audio_session.dart';
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_sound/flutter_sound.dart';
@@ -137,7 +135,7 @@ class _RecordToStreamExampleState extends State<RecordToStreamExample> {
     var recordingDataController = StreamController<Uint8List>();
     _mRecordingDataSubscription =
         recordingDataController.stream.listen((buffer) {
-      sink.add(buffer!);
+      sink.add(buffer);
     });
     await _mRecorder!.startRecorder(
       toStream: recordingDataController.sink,