+++ flutter_sound/ios/Classes/FlutterSoundRecorder.mm
@@ -83,12 +83,6 @@
  }
  
  
-- (void)closeRecorderCompleted: (bool)success
- {
-       [self invokeMethod: @"closeRecorderCompleted" boolArg: success success: success];
- 
- }
-
 
 // ------------------------------------------------------------------------------------------------------------------------------------