+++ flutter_sound_platform_interface/lib/method_channel_flutter_sound_recorder.dart
@@ -110,12 +110,6 @@ class MethodChannelFlutterSoundRecorder extends FlutterSoundRecorderPlatform {
           }
           break;
 
-        case "closeRecorderCompleted":
-          {
-            aRecorder!.closeRecorderCompleted(call.arguments['state'], success);
-          }
-          break;
-
         case "log":
           {
             int i = call.arguments['level'];