+++ flutter_sound/CHANGELOG.md
@@ -1,4 +1,4 @@
-## 9.9.0
+## 9.9.1
 
 - ## The [CHANGELOG file is here](https://flutter-sound.canardoux.xyz/changelog.html)