+++ flutter_sound_platform_interface/CHANGELOG.md
@@ -1,4 +1,4 @@
-## 9.9.0
+## 9.9.1
 
 - Please [look to this](https://flutter-sound.canardoux.xyz/changelog.html)